requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
cryptography==37.0.2
Django==3.2.8
django-cors-headers==3.10.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
ecdsa==0.17.0
greenlet==1.1.2
passlib==1.7.4
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.21
python-jose==3.3.0
pytz==2021.3
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.2.4
waitress==2.0.0
